\name{GSE20238}
\alias{GSE20238}
\docType{data}
\title{
Genes expression profiles and probesets annotation for the GPL570 subpart of the Gene Expression Omnibus GSE20238 cohort from Minguez et al., J. Hepatol 2011
}
\description{
Genes expression profiles and probesets annotation for the GPL570 subpart of the Gene Expression Omnibus GSE20238 cohort from Minguez et al., J. Hepatol 2011. Consists of a data frame GEP with the transcriptome data and another called AnnotProbeset with the annotation of probe sets.
}
\usage{data("GSE20238")}
\source{
Minguez, B., Hoshida, Y., Villanueva, A., Toffanin, S., Cabellos, L., Thung, S., Mandeli, J., Sia, D., April, C., Fan, J.-B., et al. (2011). Gene-expression signature of vascular invasion in hepatocellular carcinoma. J. Hepatol. 55, 1325-1331.
}
\references{
Minguez, B., Hoshida, Y., Villanueva, A., Toffanin, S., Cabellos, L., Thung, S., Mandeli, J., Sia, D., April, C., Fan, J.-B., et al. (2011). Gene-expression signature of vascular invasion in hepatocellular carcinoma. J. Hepatol. 55, 1325-1331.
}
\examples{
data("GSE20238")
}
\keyword{datasets}
