\name{prediction_Boyault}
\alias{prediction_Boyault}
\docType{data}
\title{
Data for two predictors from Boyault et al., Hepatology 2007
}
\description{
Data for two predictors from Boyault et al., Hepatology 2007. Consists of two lists, each concerning a classifier. The first one (prediction113symbols) contains 1 data frame containg the Gene Expression Profiles (GEP), and 2 vectors with the classes associated to the GEPs (classes) and the genes the classifier relies on. The second (prediction24probeset) contains a data frame with the data necessary, a vector of probesets and a model object.
}
\usage{data("prediction_Boyault")}
\source{
Boyault, S., Rickman, D.S., de Reynies, A., Balabaud, C., Rebouissou, S., Jeannot, E., Herault, A., Saric, J., Belghiti, J., Franco, D., et al. (2007). Transcriptome classification of HCC is related to gene alterations and to new therapeutic targets. Hepatol. Baltim. Md 45, 42-52.
}
\references{
Boyault, S., Rickman, D.S., de Reynies, A., Balabaud, C., Rebouissou, S., Jeannot, E., Herault, A., Saric, J., Belghiti, J., Franco, D., et al. (2007). Transcriptome classification of HCC is related to gene alterations and to new therapeutic targets. Hepatol. Baltim. Md 45, 42-52.
}
\examples{
data(prediction_Boyault)
}
\keyword{datasets}
