\name{prediction_Chiang}
\alias{prediction_Chiang}
\docType{data}
\title{
Genes expression profiles and clinical data for the classification from Chiang et al., Cancer Research 2008
}
\description{
Genes expression profiles and clinical data for the classification from Chiang et al., Cancer Research 2008. Consists of 1 data frame containg the Gene Expression Profiles (GEP), and 2 vectors with the classes associated to the GEPs (classes) and the genes the classifier relies on.
}
\usage{data("prediction_Chiang")}
\source{
Chiang, D.Y., Villanueva, A., Hoshida, Y., Peix, J., Newell, P., Minguez, B., LeBlanc, A.C., Donovan, D.J., Thung, S.N., Sole, M., et al. (2008). Focal gains of VEGFA and molecular classification of hepatocellular carcinoma. Cancer Res. 68, 6779-6788.
}
\references{
Chiang, D.Y., Villanueva, A., Hoshida, Y., Peix, J., Newell, P., Minguez, B., LeBlanc, A.C., Donovan, D.J., Thung, S.N., Sole, M., et al. (2008). Focal gains of VEGFA and molecular classification of hepatocellular carcinoma. Cancer Res. 68, 6779-6788.
}
\examples{
data(prediction_Chiang)
}
\keyword{datasets}
