\name{prediction_Hoshida}
\alias{prediction_Hoshida}
\docType{data}
\title{
Genes expression profiles and clinical data for the classification from Hoshida et al., Cancer Research 2009
}
\description{
Genes expression profiles and clinical data for the classification from Hoshida et al., Cancer Research 2009. Consists of 1 data frame containg the Gene Expression Profiles (GEP), and 2 vectors with the classes associated to the GEPs (classes) and the genes the classifier relies on.
}
\usage{data("prediction_Hoshida")}
\source{
Hoshida, Y., Nijman, S.M.B., Kobayashi, M., Chan, J.A., Brunet, J.-P., Chiang, D.Y., Villanueva, A., Newell, P., Ikeda, K., Hashimoto, M., et al. (2009). Integrative transcriptome analysis reveals common molecular subclasses of human hepatocellular carcinoma. Cancer Res. 69, 7385-7392.
}
\references{
Hoshida, Y., Nijman, S.M.B., Kobayashi, M., Chan, J.A., Brunet, J.-P., Chiang, D.Y., Villanueva, A., Newell, P., Ikeda, K., Hashimoto, M., et al. (2009). Integrative transcriptome analysis reveals common molecular subclasses of human hepatocellular carcinoma. Cancer Res. 69, 7385-7392.
}
\examples{
data(prediction_Hoshida)
}
\keyword{datasets}
