\name{prediction_Lee}
\alias{prediction_Lee}
\docType{data}
\title{
Genes expression profiles and clinical data for the classification from Lee et al., Hepatology 2004
}
\description{
Genes expression profiles and clinical data for the classification from Lee et al., Hepatology 2004. Consists of 2 data frames containing Gene Expression Profiles (GEP) and clinical data (ClinicalData) and a vector of the set of genes the classification relies on (Genes).
}
\usage{data("prediction_Lee")}
\source{
Lee, J.-S., Chu, I.-S., Heo, J., Calvisi, D.F., Sun, Z., Roskams, T., Durnez, A., Demetris, A.J., and Thorgeirsson, S.S. (2004). Classification and prediction of survival in hepatocellular carcinoma by gene expression profiling. Hepatol. Baltim. Md 40, 667-676.
}
\references{
Lee, J.-S., Chu, I.-S., Heo, J., Calvisi, D.F., Sun, Z., Roskams, T., Durnez, A., Demetris, A.J., and Thorgeirsson, S.S. (2004). Classification and prediction of survival in hepatocellular carcinoma by gene expression profiling. Hepatol. Baltim. Md 40, 667-676.
}
\examples{
data(prediction_Lee)
}
\keyword{datasets}
