\name{prediction_Roessler}
\alias{prediction_Roessler}
\docType{data}
\title{
Genes expression profiles and clinical data for the classification from Roessler et al., Cancer Research 2010
}
\description{
Genes expression profiles and clinical data for the classification from Roessler et al., Cancer Research 2010. Consists of 2 data frames containing Gene Expression Profiles (GEP) and clinical data (ClinicalData) and a vector of the set of genes the classification relies on (Genes).
}
\usage{data("prediction_Roessler")}
\source{
Roessler, S., Jia, H.-L., Budhu, A., Forgues, M., Ye, Q.-H., Lee, J.-S., Thorgeirsson, S.S., Sun, Z., Tang, Z.-Y., Qin, L.-X., et al. (2010). A unique metastasis gene signature enables prediction of tumor relapse in early-stage hepatocellular carcinoma patients. Cancer Res. 70, 10202-10212.
}
\references{
Roessler, S., Jia, H.-L., Budhu, A., Forgues, M., Ye, Q.-H., Lee, J.-S., Thorgeirsson, S.S., Sun, Z., Tang, Z.-Y., Qin, L.-X., et al. (2010). A unique metastasis gene signature enables prediction of tumor relapse in early-stage hepatocellular carcinoma patients. Cancer Res. 70, 10202-10212.
}
\examples{
data(prediction_Roessler)
}
\keyword{datasets}
