\name{prognosis_Nault}
\alias{prognosis_Nault}
\docType{data}
\title{
Gene expression and clinical data used for the prognosis classification from Nault and Zucman Rossi, International Journal of Hepatology 2013
}
\description{
Gene expression and clinical data used for the prognosis classification from Nault and Zucman Rossi, International Journal of Hepatology 2013. Consists of 2 data frames containing Gene Expression Profiles (GEP) and clinical data (ClinicalData).
}
\usage{data("prognosis_Nault")}
\source{
Nault, J.C., and Zucman Rossi, J. (2013). Molecular classification of hepatocellular adenomas. Int. J. Hepatol. 2013, 315947.
}
\references{
Nault, J.C., and Zucman Rossi, J. (2013). Molecular classification of hepatocellular adenomas. Int. J. Hepatol. 2013, 315947.
}
\examples{
data(prognosis_Nault)
}
\keyword{datasets}
